Swap okio for kotlinx-io across solana-kotlin internals (#299)
okio.Buffer/BufferedSource are replaced with kotlinx.io.Buffer/Source
across the 14 commonMain files that used them, plus the codegen
templates that emit binary serialisation code.

Public API surface stays free of kotlinx-io types. Three previously-
public deserialisation helpers (FeeCalculator.read, NonceAccountData.read,
PublicKey.Companion.read) accepted okio.BufferedSource, so they were
hidden as `internal` rather than re-exposed under a kotlinx.io.Source
parameter. They were already used only inside this module.

Codegen template changes (codegen/.../ProgramGenerator.kt):
- Buffer().chain().chain().readByteArray() pattern rewritten to
  Buffer().apply { ... }.readByteArray() because kotlinx-io's write
  methods return Unit
- writeByte(Int) -> writeByte(Byte): every emitted call now wraps with
  .toByte() / .toShort() as needed
- writeUtf8 -> writeString
- New imports: kotlinx.io.{Buffer, readByteArray, writeIntLe, writeLongLe,
  writeShortLe, writeString}
- sizePrefixTypeNode for strings now uses ${paramName}Bytes to avoid
  variable name collisions when an instruction has multiple string fields

Generated files in solana-kotlin/src/commonMain/generated/ regenerated
to match.

okio is removed from libs.versions.toml; kotlinx-io 0.9.0 takes its
place. BCV baseline updated for the now-internal read helpers.

+++ solana-kotlin/src/commonMain/kotlin/net/avianlabs/solana/domain/core/SignedTransaction.kt
@@ -3,9 +3,10 @@ package net.avianlabs.solana.domain.core
 import co.touchlab.skie.configuration.annotations.SkieVisibility
 import net.avianlabs.solana.tweetnacl.TweetNaCl
 import net.avianlabs.solana.tweetnacl.ed25519.PublicKey
+import kotlinx.io.Buffer
+import kotlinx.io.readByteArray
 import net.avianlabs.solana.tweetnacl.vendor.encodeToBase58String
 import net.avianlabs.solana.vendor.ShortVecEncoding
-import okio.Buffer
 
 @Deprecated(
   message = "Use VersionedTransaction instead",
@@ -87,6 +88,6 @@ public data class SignedTransaction(
     out.write(signaturesLength)
     orderedSigs.forEach(out::write)
     out.write(serializedMessage)
-    return SerializedTransaction(out.readByteArray(bufferSize.toLong()))
+    return SerializedTransaction(out.readByteArray(bufferSize))
   }
 }
